feat(IMDbSearch): add new language detection API
A new API (https://github.com/zappbrandigan/lang-detect.git) replaced the npm npm language detection
package that was being used previously. This fixes the issues with accuracy in languages such as
Portuguese and French.

This adds an addition API call. So, the AKA titles have been separated from the production details
so that they load separately.